README: clarify RickLink instructions on lpc55s69xpresso
* Overwrite existing firmware so the new bootloader doesn't accidentally
try to boot an old main firmware
* Remove instruction to setup three terminals (pyocd gdbserver, telnet
for semihosting, gdb) as multiple folks have run into problems following
that bit.  Once RickLink and pyocd is installed, `cargo xtask flash` and
`humility itm` just work without running a gdb server.  Semihosting
isn't enabled in the lpc55s69xpresso app so no need to run a terminal
for it.  Running a pyocd gdbserver is necessary if you want to run
`cargo xtask gdb` but we don't tell people to do that.